pyproject.toml
[project]
name = "ai4eu_server_demo"
description = "A containerized demo application for AI on demand (previously AI4EU) "
version = "1.0.20230911"
requires-python = ">=3.11"
authors = [

    { name = "Antonis Ganios"},
    { name = "Íñigo Aréjula Aísa"},
    { name = "Jos van der Velde", email = "j.d.v.d.velde@tue.nl"},
    { name = "Pieter Gijsbers", email = "p.gijsbers@tue.nl"},
    { name = "Giannis Mouchakis"},
    {name = "Taniya Das", email = "t.das@tue.nl"}
]
dependencies = [
    "urllib3== 2.0.7",
    "bibtexparser==1.4.1",
    "datasets==2.14.6",
    "fastapi==0.104.1",
    "uvicorn==0.24.0.post1",
    "requests==2.31.0",
    "mysqlclient==2.2.0",
    "oic==1.6.0",
    "python-keycloak==3.3.0",
    "python-dotenv==1.0.0",
    "pydantic_schemaorg==1.0.6",
    "python-dateutil==2.8.2",
    "sqlmodel==0.0.11",
    "httpx==0.25.1",
    "sickle==0.7.0",
    "xmltodict==0.13.0",
    "python-multipart==0.0.6",
    "mysql-connector-python==8.2.0",
    "elasticsearch==8.10.1",
]
readme = "README.md"

[project.optional-dependencies]
dev = [
    "types-python-dateutil==2.8.19.14",
    "pytest==7.4.3",
    "pytest-dotenv==0.5.2",
    "pytest-xdist==3.4.0",
    "pre-commit==3.5.0",
    "responses==0.24.0",
    "starlette==0.27.0"
]

[tool.setuptools]
py-modules = []

[tool.black]
line-length = 100

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::FutureWarning"
]
env_override_existing_values = 1
env_files = [
    "src/.env",
    "src/tests/.env"
]



